Calibrate the championship projection's range
The projection held every team's pace fixed at whatever the ranker said that
weekend, so the only thing moving a simulated season was race-to-race noise -
and noise averages out over a dozen races. The published 10th-90th band was
correspondingly narrow: graded against real final constructors' standings it
contained the truth 47% of the time against a stated 80%.

Development was the obvious suspect and measurably not the cause. Fitting the
change in a team's mean finishing position either side of a checkpoint, with
the sampling noise in both means modelled rather than counted as development,
puts it at 1.37 positions over a full season (95% 0.90-1.83, 283 team-
checkpoints). Adding exactly that much moved coverage by two points.

What does not average out is the model being wrong about a car today and
carrying that into every remaining race. Each simulated season now draws one
pace offset per team, sized on 2019-2022 and graded on 2023-2025 which the
sweep never saw: 47% to 71%. Still short of 80%, and the method page says so.

+++ src/f1pred/title_backtest.py
@@ -65,7 +65,7 @@ def run(
             continue
 
         for frac in CHECKPOINT_FRACTIONS:
-            after = int(round(len(rounds) * frac))
+            after = round(len(rounds) * frac)
             if after < 3 or after >= len(rounds):
                 continue
             nxt = df[(df.season == season) & (df["round"] == after + 1)]
